requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
Pillow==9.5.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.7.0
pytest==8.3.4
pytest-django==4.9.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.2.1
typing_extensions==4.6.3
